+++ build.gradle
@@ -25,7 +25,7 @@ ext {
 apply plugin: 'application'
 apply from: 'gradle/formatting.gradle'
 
-mainClassName = 'transportservice.ExtensionsRunner'
+mainClassName = 'opensearchSDK.transport.ExtensionsRunner'
 
 group 'org.example'
 version '1.0-SNAPSHOT'
