+++ src/main/java/opensearchSDK/transport/CopyBytesServerSocketChannel.java
@@ -43,7 +43,7 @@
  * Modifications Copyright OpenSearch Contributors. See
  * GitHub history for details.
  */
-package transportservice;
+package opensearchSDK.transport;
 
 import io.netty.channel.socket.nio.NioServerSocketChannel;
 import io.netty.util.internal.SocketUtils;
