+++ src/test/java/TestNetty4Transport.java
@@ -18,8 +18,8 @@ import org.opensearch.test.OpenSearchTestCase;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.TransportSettings;
 
-import transportservice.ExtensionsRunner;
-import transportservice.netty4.Netty4Transport;
+import opensearchSDK.netty4.Netty4Transport;
+import opensearchSDK.transport.ExtensionsRunner;
 
 public class TestNetty4Transport extends OpenSearchTestCase {
